minor changes before showing to oisin

+++ pyTools/misc/basic.py
@@ -75,7 +75,7 @@ def countInt(x, minLength=None):
     m = np.min(x)
     if m > 0:
         m = 0
-    cnt = np.bincount(x - m, minlength=minLength)
+    cnt = np.bincount((x - m).astype(np.int), minlength=minLength)
     cntMat = np.asarray(list(enumerate(cnt)))
     cntMat[:,0] += m
     return cntMat
